btrfs-progs: add option "--time-format=short|iso|unix|locale" to subvolume list
Affects time format of "otime". Supporting unix time and ISO8601 makes
the output of "subvolume list" machine-readable.

Default is "short", keeping default output as it was before (I suggest
changing the default to "iso", as with "short" is it not clear to the
user if localtime or gmtime is printed, and it does not contain
whitespace).

Signed-off-by: Axel Burri <axel@tty0.ch>